+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Minor: Consolidate font picking into one dialog. (#6531)
 - Minor: Added a menu action to sort tabs alphabetically. (#6551)
 - Minor: Fixed "edit hotkey" dialog opening like a normal window. (#6540)
+- Minor: Added "Open in logs" message context action
 - Minor: Change Bestlogs redirect to Supatv logs
 - Bugfix: Expose the "Extra extension IDs" setting on non-Windows systems too. (#6509)
 - Bugfix: Fixed some commands and filters not working as expected in seach popups. (#6539)